requirements.txt
docopt==0.6.2
pytest-cov
